Added new processor handler object
The new ProcessorHandler class is used to manage
all of the classes derived from the Processor
parent class. When the map file is loaded, individual
processors are added to the handler (if they exist).
When the unpacker calls for the processing of the
raw event, each of the channel signals in the raw
event will be given to their respective processor
for processing.