Added detector location to map entry
Added location tracking to MapEntry. Detectors of
the same class are now given an id based on their
order in the map file. The first instance of a
detector type has a location of one.

Added root output to VandleProcessor.

Began implementation of event pulse fitting by
adding new methods to Processor class.