Added high resolution timing infrastructure
Added a lot of infrastructure for doing high-res
timing on trace captures.

Unpacker::ReadSpill was changed to accept an
unsigned int array instead of a char array. This
was done to prevent possible runtime errors when
converting from char* to unsigned int* which is
NOT guaranteed to work in c++;

Added new GenericProcessor.cpp for use with
detector types which do not fit any other category.