Added ability to specify beta and gamma
Processors now operate using ChannelEventPair class
which contains a pointer to a ChannelEvent and a
MapEntry. This was done so that processors will
have access to the map entry for a given event.

Added beta and gamma fitting parameters to MapEntry.
These values may now be specified through the input
map file.

Added *.dat files to the .gitignore.