Fixed max module check error
Fixed an error in Mapfile class which was incorrectly
checking the read channel number against the max
number of modules.

Added debug information to warnings from MapFile.

Changed the location of defined detectors from their
position in the map file to (100*module + channel)
so that the data in the output file will not be
dependent on the input map file.