Changed the way that Processor::PreProcess works
The PreProcess method of Processor was modified in
order to make overloading of ::FitPulses simpler.

PreProcess now performs the majority of tasks which
FitPulses did previously.

FitPulses has been renamed to FitPulse and now only
fits one pulse (in the form of a TGraph) at a time.

Removed from un-needed includes to speed up compile
time.