Added PhoswichProcessor
Added a new processor for phoswich detectors. This
processor uses a special fitting routine to fit the
fast and the slow components of the pulse
individually.

Added phoswich data structure to def.struct.

Added TFitResultPtr to Processor class so that
processors may access fit information.

Added .def.struct.bkp to the .gitignore.