Fixed processor events not clearing correctly
Fixed a bug which prevented processors from clearing
their channel event queues when no start event was
found within a time window. This caused all processor
events found in a file to be added to the output tree
regardless of whether or not a valid start signal was
found for that time window.