Added ability to select various histograms
Added a vector of TH1* to OnlineProcessor so that
the user may choose from any number of histograms
which have been added to the histogram list.

Added GetHists method to Processor class which may
be used to add processor-specific histograms to the
online processor for plotting.

Added raw data histograms to the plotting list as
default plots.