Implemented CFD analysis routine
Implemented a new CFD analysis routine to the class
Processor. This routine was added in order to have
a fast alternative to slow fitting while preserving
the ability to calculate high-res timing.

CFD parameters may be passed to individual detector
processors through the map file in the same way that
beta and gamma may be passed for fitting.

Root fitting is no longer enabled by default. It may
be re-enabled by passing the '--fitting' flag at the
command line.