Added ability to change configuration directory
Added the ability to change where the makefile
installs the runtime configuration files. This
allows the user to install the config files to
a directory other than SimplePixieScan. This
prevents cluttering up the build directory.

The makefile will unpack whichever config files
are not already found in the config directory.
This means that the scan code will still run
even if the user is missing some of the required
files in the config directory they specified, but
their existing files will not be overwritten.