Fixed reading of data from presorted files
Reading of data from presorted data files now works
as expected.

Testing on Cf data for VANDLE yields the same ctof
plot for presorted data as it does for scanning from
the original data file except that the presorted one
contains one extra event. This needs to be tested
further.