Set beta and gamma to fixed in fitting routine
I changed SetParameter to FixParameter for params
beta and gamma in the fitting routine. This was
done to increase the speed of the scan (when fitting
is enabled) and to increase the reliability of the
fit results.

It seems that beta and gamma are too sensitive to be
left variable.