Updated double-ended event handler
Updated the processor method which handles events
for double-ended detectors. Changed the way that
double-ended events are sorted. They are no longer
sorted only by channel ID. Instead, they are now
sorted into pairs. This should reduce the number
of events which are invalid because neighboring
channels are not found next to each other.

Added much more lost event tracking to Processor
class. Events which are missing from the output
tree are now tracked and an explanation of the
number of missing events is now given when each
processor displays the total number of processed
events.