Added better branch error checking to cmbinner
Modified TQDC bin widths for TQDC histograms. They
now use the empirical formula for calculating the
TQDC (in keVee) from the energy of the neutron for
setting the width of the bins. The result is that
smaller TQDC values will have smaller bin widths
to match those of the variable width energy histos.