Added segmented PSPMT processor type
Unlike GenericBar and LiquidBar, PSPmtProcessor will
handle both single and double-ended detectors.

To process a PSPMT, add something like

1 0 pspmt::
1 1 pspmt::SE
1 2 pspmt::NE
1 3 pspmt::NW
1 4 pspmt::SW

to your map.dat file. To switch to a PSPMT bar,
add either "left" or "right" to the detector
subtype. The four corners are defined using the
four intercardinal directions SE, NE, NW, and SW
as the detector tag. The dynode signal is assumed
to be the channel with no direction in the detector
tag.