Added high-level PSPMT analyzer
The PSPMT analyzer program "pspmt" operates in a
similar way to "barifier". The primary difference
is that pspmt requires an additional config file
named "pspmt.dat" which acts as the pspmt map. The
format of this map file is as follows

example "pspmt.dat":
  # DYNODE SE NE NW SW
  4 6 8 10 14
  5 15 11 9 7

In the example above, the first PSPMT has a dynode
in channel 4 and anodes in channels 6, 8, 10, and 14.
The second has dynode in 5 and anodes in 15, 11, 9,
and 7.