Modified barifier uncertainty smearing
Modified the way barifier computes the random point
inside of the detector. Instead of adding a small
random angle to the center in cylindrical coords,
barifier will now select a random x and y inside
the detector and transform it to the lab frame.
This is more in line with how the pspmt analyzer
works and will allow much more complicated detector
geometries.

Added barifier option which allows the use of the
left/right light balance to compute the "vertical"
position (y) instead of the time difference. This
option works better for shorter detectors.