Modified simpleScan online-mode range commands
simpleScan online-mode 'range' commands now set
the range of the histogram currently drawn on a
pad instead of the histogram at a certain index.
This is now in line with how the x(y,z)log cmds
work.

simpleScan online-mode help syntax modified to
better reflect the function of user input variables.