Modified the way pspmt map entries are decoded
Added new class MapEntryValidator which is used
by the MapFile class to validate entries from the
map when searching for entries which match certain
criteria.

The updated map search fixes pspmt detectors not
being initialized correctly when in --untriggered
mode or when one of the pspmt signals is used as
the start signal.